Move FakeStatsdServer into s.m.m.testing.
