+++ docs/api.rst
@@ -20,3 +20,13 @@ Statsd Implementation
 ---------------------
 .. autoclass:: sprockets.mixins.metrics.StatsdMixin
    :members:
+
+Testing Helpers
+---------------
+*So who actually tests that their metrics are emitted as they expect?*
+
+Usually the answer is *no one*.  Why is that?  The ``testing`` module
+contains some helper that make testing a little easier.
+
+.. autoclass:: sprockets.mixins.metrics.testing.FakeStatsdServer
+   :members:
